Bug fix, WIP tests
